+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Validations - Start the support of forms validations (with 5 first validations).
 
 ## RELEASE 1.1.12 - 2017-07-11
 ### Added